# Servers/swagger.yaml
openapi: 3.0.0
info:
  title: VerifyWise API
  description: AI Governance Platform API
  version: 2.0.0
servers:
  - url: /api
    description: Main API server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /ai-detection/repositories:
    get:
      tags:
        - AI Detection
      summary: List Repositories
      operationId: listRepositories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
    post:
      tags:
        - AI Detection
      summary: Create Repository
      operationId: createRepository
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created successfully
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
  /ai-detection/repositories/{id}:
    get:
      tags:
        - AI Detection
      summary: Get Repository
      operationId: getRepository
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
    patch:
      tags:
        - AI Detection
      summary: Update Repository
      operationId: updateRepository
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
    delete:
      tags:
        - AI Detection
      summary: Delete Repository
      operationId: deleteRepository
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
  /ai-detection/repositories/{id}/scan:
    post:
      tags:
        - AI Detection
      summary: Trigger Repository Scan
      operationId: triggerRepositoryScan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created successfully
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
  /ai-detection/repositories/{id}/scans:
    get:
      tags:
        - AI Detection
      summary: Get Repository Scans
      operationId: getRepositoryScans
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
  /ai-detection/repositories/{id}/webhook-secret:
    post:
      tags:
        - AI Detection
      summary: Generate Webhook Secret Controller
      operationId: generateWebhookSecretController
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created successfully
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
...